backend/requirements.txt
google-adk>=1.15.0
fastapi
uvicorn
pydantic